+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Make all internal modules `internal`
 
 ## 1.0.0 - 2020-10-27
 - Initial implementation